Break up parser.py into many files
